refactor(query): Complete Phase 2.2 - Extract major phenotypic filter callbacks
Successfully extracted all major phenotypic filter management callbacks from the
monolithic pages/query.py to the modular query/callbacks/filters.py architecture.

Major callbacks extracted:
- manage_phenotypic_filters() - Complex state management for phenotypic filters (160+ lines)
- render_phenotypic_filters() - UI rendering for phenotypic filters (180+ lines)
- update_live_participant_count() - Live participant count updates (103+ lines)
- update_phenotypic_session_notice() - Session notice updates

Impact:
- Reduced pages/query.py from ~1385 to 932 lines (-453 lines, 33% reduction)
- Added 664 lines to modular query/callbacks/filters.py
- All callbacks properly registered and tested for import compatibility
- Maintained full backward compatibility and functionality

This completes Phase 2.2 of the Query Refactoring Plan, successfully modularizing
the most complex filtering functionality while preserving all existing behavior.
